Changed seeding position to uint32_t because I had like 200K

+++ apps/rosetta/riflib/rifdock_typedefs.hh
@@ -107,7 +107,7 @@ typedef typename ScaffoldProvider::ScaffoldIndex ScaffoldIndex;
 
 struct RifDockIndex {
     uint64_t nest_index;
-    uint16_t seeding_index;
+    uint32_t seeding_index;
     ScaffoldIndex scaffold_index;
 
     RifDockIndex() :
